modificacion de servicios

+++ src/main/java/com/business/backendBusiness/Repository/WorkRepository.java
@@ -3,5 +3,5 @@ package com.business.backendBusiness.Repository;
 import com.business.backendBusiness.entity.Work;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkRepository extends JpaRepository <Work,Integer> {
+public interface WorkRepository  extends JpaRepository<Work,Long> {
 }
